aioa/pyproject.toml
[project]
name = "aio-analyzer"
version = "0.1.0"
description = "AI & Search Visibility Benchmarking Tool — measures competitive positioning across LLMs and search engines"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "openai>=1.0",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "aiosqlite>=0.20",
    "mcp>=1.0",
]

[project.scripts]
aio-analyzer = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.7",
]

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
select = ["I", "E", "W", "F", "UP", "B", "SIM", "C4"]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["S101", "E501"]
"src/orchestrator.py" = ["E501"]
"src/output.py" = ["E501"]
"src/stores/sqlite_store.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
